# Conflicts:
#	FitCloudKitDemo/FitCloudKitDemo.xcodeproj/project.pbxproj
#	FitCloudKitDemo/FitCloudKitDemo/AppDelegate.m
#	FitCloudKitDemo/FitCloudKitDemo/DemoListController.m
#	FitCloudKitDemo/FitCloudKitDemo/ViewController.m
#	FitCloudKitDemo/Podfile.lock
#	TopStepComKit/OtherSDK/flywearsdk.framework/_CodeSignature/CodeSignature
#	TopStepComKit/OtherSDK/h264encoder.framework/Info.plist
#	TopStepComKit/Products/TopStepComKit.framework/Frameworks/flywearsdk.framework/_CodeSignature/CodeResources
#	TopStepComKit/Products/TopStepComKit.framework/Frameworks/flywearsdk.framework/flywearsdk
#	TopStepComKit/Products/TopStepComKit.framework/Frameworks/persimwearSDK.framework/persimwearSDK
#	TopStepComKit/Products/TopStepComKit.framework/Modules/TopStepComKit.swiftmodule/Project/arm64-apple-ios.swiftsourceinfo
#	TopStepComKit/Products/TopStepComKit.framework/Modules/TopStepComKit.swiftmodule/arm64-apple-ios.swiftmodule
#	TopStepComKit/Products/TopStepComKit.framework/TopStepComKit
#	TopStepComKit/Products/TopStepComKit.framework/_CodeSignature/CodeDirectory
#	TopStepComKit/Products/TopStepComKit.framework/_CodeSignature/CodeRequirements-1
#	TopStepComKit/Products/TopStepComKit.framework/_CodeSignature/CodeResources
#	TopStepComKit/Products/TopStepComKit.framework/_CodeSignature/CodeSignature
#	TopStepComKit/fw/flywearsdk.framework/_CodeSignature/CodeDirectory
#	TopStepComKit/fw/flywearsdk.framework/_CodeSignature/CodeRequirements-1
#	TopStepComKit/fw/flywearsdk.framework/flywearsdk

+++ TopStepComKit/Products/TopStepComKit.framework/Headers/TPSHistoryActivityModel.h
@@ -19,7 +19,7 @@
 @property (nonatomic, assign) int stand;
 
 @property (nonatomic, assign) int steps;
-@property (nonatomic, assign) int sport;//单位分钟，运动时长
+@property (nonatomic, assign) double sport;//单位分钟，运动时长
 @property (nonatomic, assign) int number;//表示这个时间段内的活动次数
 @property (nonatomic, assign) double calorie;//表示这个时间段内活动消耗的卡路里，单位千卡
 @property (nonatomic, assign) double distance;//表示这个时间段内活动的距离，单位米
